rename core.core to core.concept
