+++ test/core/test_config.py
@@ -8,7 +8,7 @@ class TestConfig1(TestCase):
 
     def test_exported_variable(self):
         self.assertEqual(
-            siibra.core.core.GITLAB_PROJECT_TAG,
+            siibra.core.concept.GITLAB_PROJECT_TAG,
             f"siibra-{siibra.__version__}"
         )
 
@@ -16,8 +16,8 @@ class TestConfig1(TestCase):
     def test_when_env_set(self):
 
         importlib.reload(siibra)
-        importlib.reload(siibra.core.core)
-        self.assertEqual(siibra.core.core.GITLAB_PROJECT_TAG,"develop")
+        importlib.reload(siibra.core.concept)
+        self.assertEqual(siibra.core.concept.GITLAB_PROJECT_TAG,"develop")
 
 if __name__ == "__main__":
     unittest.main()
